refactor(commands): extract doodles commands
