+++ src/commands/index.js
@@ -11,6 +11,10 @@ const delmod = require('./mods/delmod');
 const aliases = require('./aliases/aliases');
 const addalias = require('./aliases/addalias');
 const delalias = require('./aliases/delalias');
+const doodles = require('./doodles/doodles');
+const track = require('./doodles/track');
+const untrack = require('./doodles/untrack');
+const report = require('./doodles/report');
 
 const commands = new Discord.Collection();
 
@@ -24,6 +28,10 @@ commands.set(delmod.name, delmod);
 commands.set(aliases.name, aliases);
 commands.set(addalias.name, addalias);
 commands.set(delalias.name, delalias);
+commands.set(doodles.name, doodles);
+commands.set(track.name, track);
+commands.set(untrack.name, untrack);
+commands.set(report.name, report);
 
 // Generate help command from command definitions
 const help = helpGen(commands);
